Extract Sorbet related services from `symbol_loader`
