+++ lib/tapioca.rb
@@ -35,6 +35,7 @@ require "tapioca/config_builder"
 require "tapioca/generator"
 require "tapioca/cli"
 require "tapioca/gemfile"
+require "tapioca/compilers/sorbet"
 require "tapioca/compilers/symbol_table_compiler"
 require "tapioca/compilers/symbol_table/symbol_generator"
 require "tapioca/compilers/symbol_table/symbol_loader"
